NN-1560 Reorganised some pay related files and began tests for PayOptions
