+++ src/Components/ModalContainer/index.js
@@ -1,0 +1,3 @@
+import ModalContainer from './ModalContainer'
+
+export default ModalContainer
