--- src/Components/PayOffender/index.js
@@ -1,3 +1,0 @@
-import PayOffender from './PayOffender'
-
-export default PayOffender
