+++ src/Components/PayOtherForm/index.js
@@ -1,0 +1,3 @@
+import PayOtherForm from './PayOtherForm'
+
+export default PayOtherForm
