+++ src/ResultsActivity/elements/PayOptions/index.js
@@ -1,0 +1,3 @@
+import PayOptions from './PayOptions'
+
+export default PayOptions
